Moving user routes into /user and simplifying top level router in app.js
