+++ app/app.js
@@ -1,7 +1,5 @@
 const eventRouter = require("./routes/event");
-const signupRouter = require("./routes/signup");
-const loginRouter = require("./routes/login");
-const logoutRouter = require("./routes/logout");
+const userRouter = require("./routes/user");
 
 //Prepare appliation with middleware
 const getAppWithMiddleware = require("./middleware");
@@ -13,9 +11,7 @@ app.use("/api/event", eventRouter.api);
 app.use("/admin/event", eventRouter.adminView);
 app.use("/api/admin/event", eventRouter.adminApi);
 
-app.use("/signup", signupRouter);
-app.use("/login", loginRouter);
-app.use("/logout", logoutRouter);
+app.use("/user", userRouter);
 
 app.get("/", (req, res) => res.send("Hello World!"));
