+++ app/routes/user/logout.js
